+++ aqua/gribber/gribber.py
@@ -7,7 +7,7 @@ import subprocess
 from glob import glob
 from aqua.logger import log_configure
 from aqua.util import load_yaml, dump_yaml, create_folder
-from aqua.util import get_config_dir, get_machine
+from aqua.util import ConfigPath
 from aqua.reader import Reader
 
 
@@ -103,11 +103,9 @@ class Gribber():
                                     self.exp)
         self.logger.info("JSON directory: %s", self.jsondir)
 
-        if not self.dir['configdir']:
-            self.configdir = get_config_dir()
-        else:
-            self.configdir = self.dir['configdir']
-        self.machine = get_machine(self.configdir)
+        Configurer = ConfigPath(self.dir['configdir'])
+        self.configdir = Configurer.configdir
+        self.machine = Configurer.machine
 
         self.logger.info("Data directory: %s", self.datadir)
         self.logger.info("JSON directory: %s", self.jsondir)
@@ -121,7 +119,7 @@ class Gribber():
                              self.tgt_json)
         else:
             self.tgt_json = self.source
-            self.logger.info("json file will be named as source.",
+            self.logger.info("json file will be named as %s.",
                              self.source)
         self.indices = None
 
@@ -137,9 +135,10 @@ class Gribber():
             self.gribfiles = '*'+format
 
         # Get catalog filename
-        self.catalogfile = os.path.join(self.configdir, 'machines',
-                                        self.machine, 'catalog',
-                                        self.model, self.exp+'.yaml')
+        self.catalogdir = os.path.join(self.configdir, 'machines',
+                                       self.machine, 'catalog',
+                                       self.model)
+        self.catalogfile = os.path.join(self.catalogdir, self.exp+'.yaml')
         self.logger.warning("Catalog file: %s", self.catalogfile)
 
         # Get JSON filename
@@ -169,11 +168,16 @@ class Gribber():
         # Create JSON file
         if self.flag[1]:
             self._create_json()
+            if not os.path.exists(self.jsonfile):
+                self.logger.error("Gribscan has created different json filename!")
+                self.logger.error("Please check and modify the catalog files accordingly")
 
         # Create catalog entry
         self._create_catalog_entry()
 
         self._create_main_catalog()
+
+        self._create_regrid_entry()
 
     def check_entry(self):
         """
@@ -181,9 +185,9 @@ class Gribber():
         """
         self.reader = Reader(model=self.model, exp=self.exp,
                              source=self.source, configdir=self.configdir,
-                             loglevel=self.logger.level)
+                             loglevel=self.logger.level,fix=False)
 
-        data = self.reader.retrieve(fix=False)
+        data = self.reader.retrieve()
         assert len(data) > 0
 
     def _check_steps(self):
@@ -371,6 +375,9 @@ class Gribber():
             cat_file['sources'] = {}
             cat_file['sources'][self.source] = block_cat
 
+            # Create folder if needed
+            create_folder(folder=self.catalogdir)
+
         # Write catalog file
         dump_yaml(outfile=self.catalogfile, cfg=cat_file)
 
@@ -411,10 +418,19 @@ class Gribber():
                     in {mainfilepath}. Skipping...")
             return
         else:  # Source does not exist
-            main_file['sources'][self.source] = block_main
+            main_file['sources'] = block_main
 
         # Write catalog file
         dump_yaml(outfile=mainfilepath, cfg=main_file)
+
+    def _create_regrid_entry(self):
+        """
+        Updates the regrid.yaml file.
+
+        At the moment this is only a placeholder for a possible function.
+        """
+
+        self.logger.warning("Update regrid.yaml has to be done manually.")
 
     def help(self):
         """