+++ aqua/reader/reader.py
@@ -1,10 +1,10 @@
 """The main AQUA Reader class"""
 
 import os
-import sys
 import re
 
 import types
+import tempfile
 import intake
 import intake_esm
 
@@ -15,9 +15,8 @@ import numpy as np
 import smmregrid as rg
 
 from aqua.util import load_yaml, load_multi_yaml
-from aqua.util import get_reader_filenames, get_config_dir, get_machine
-from aqua.util import log_history, log_history_iter
-from aqua.logger import log_configure
+from aqua.util import ConfigPath, area_selection
+from aqua.logger import log_configure, log_history, log_history_iter
 import aqua.gsv
 
 from .streaming import Streaming
@@ -27,8 +26,9 @@ from .reader_utils import check_catalog_source, group_shared_dims, set_attrs
 
 
 # default spatial dimensions and vertical coordinates
-default_space_dims = ['i', 'j', 'x', 'y', 'lon', 'lat', 'longitude', 'latitude',
-                      'cell', 'cells', 'ncells', 'values', 'value', 'nod2', 'pix', 'elem']
+default_space_dims = ['i', 'j', 'x', 'y', 'lon', 'lat', 'longitude',
+                      'latitude', 'cell', 'cells', 'ncells', 'values',
+                      'value', 'nod2', 'pix', 'elem']
 
 
 # set default options for xarray
@@ -38,13 +38,15 @@ xr.set_options(keep_attrs=True)
 class Reader(FixerMixin, RegridMixin):
     """General reader for NextGEMS data."""
 
-    def __init__(self, model="ICON", exp="tco2559-ng5", source=None, freq=None,
+    def __init__(self, model=None, exp=None, source=None, freq=None, fix=True,
                  regrid=None, method="ycon", zoom=None, configdir=None,
                  areas=True,  # pylint: disable=W0622
                  datamodel=None, streaming=False, stream_step=1, stream_unit='steps',
-                 stream_startdate=None, rebuild=False, loglevel=None, nproc=4):
+                 stream_startdate=None, rebuild=False, loglevel=None, nproc=4, aggregation=None, verbose=False,
+                 buffer=None):
         """
-        Initializes the Reader class, which uses the catalog `config/config.yaml` to identify the required data.
+        Initializes the Reader class, which uses the catalog
+        `config/config.yaml` to identify the required data.
 
         Args:
             model (str, optional): Model ID. Defaults to "ICON".
@@ -52,7 +54,8 @@ class Reader(FixerMixin, RegridMixin):
             source (str, optional): Source ID. Defaults to None.
             regrid (str, optional): Perform regridding to grid `regrid`, as defined in `config/regrid.yaml`. Defaults to None.
             method (str, optional): Regridding method. Defaults to "ycon".
-            zoom (int):             healpix zoom level. (Default: None)
+            fix (bool, optional): Activate data fixing
+            zoom (int): healpix zoom level. (Default: None)
             configdir (str, optional): Folder where the config/catalog files are located. Defaults to None.
             areas (bool, optional): Compute pixel areas if needed. Defaults to True.
             var (str or list, optional): Variable(s) to extract; "vars" is a synonym. Defaults to None.
@@ -64,6 +67,9 @@ class Reader(FixerMixin, RegridMixin):
             rebuild (bool, optional): Force rebuilding of area and weight files. Defaults to False.
             loglevel (str, optional): Level of logging according to logging module. Defaults to log_level_default of loglevel().
             nproc (int,optional): Number of processes to use for weights generation. Defaults to 16.
+            aggregation (str, optional): aggregation to be used for GSV access (one of S (step), 10M, 15M, 30M, 1H, H, 3H, 6H, D, W, M, Y). Defaults to None (using default from catalogue).
+            verbose (bool, optional): if to print to screen additional info (used only for FDB access at the moment)
+            buffer (str, optional): buffering of FDB/GSV streams in a temporary directory specified by the keyword. The result will be a dask array and not an iterator.
 
         Returns:
             Reader: A `Reader` class object.
@@ -79,6 +85,8 @@ class Reader(FixerMixin, RegridMixin):
         self.freq = freq
         self.vert_coord = None
         self.deltat = 1
+        self.aggregation = aggregation
+        self.verbose = verbose
         extra = []
 
         self.grid_area = None
@@ -90,76 +98,95 @@ class Reader(FixerMixin, RegridMixin):
                                   stream_unit=stream_unit,
                                   stream_startdate=stream_startdate,
                                   loglevel=loglevel)
-        # Export streaming methods
+
+        # Export streaming methods TO DO: probably useless
         self.reset_stream = self.streamer.reset_stream
         self.stream = self.streamer.stream
         self.stream_generator = self.streamer.stream_generator
 
-        if not configdir:
-            self.configdir = get_config_dir()
+        self.previous_data = None  # used for FDB iterator fixing 
+        if buffer:  # optional FDB buffering
+            if not os. path. isdir(buffer):
+                raise ValueError("The directory specified by buffer must exist.") 
+            self.buffer = tempfile.TemporaryDirectory(dir=buffer)
         else:
-            self.configdir = configdir
-        self.machine = get_machine(self.configdir)
+            self.buffer = None
+
+        # define configuration file and paths
+        Configurer = ConfigPath(configdir=configdir)
+        self.configdir = Configurer.configdir
+        self.machine = Configurer.machine
 
         # get configuration from the machine
         self.catalog_file, self.regrid_file, self.fixer_folder, self.config_file = (
-            get_reader_filenames(self.configdir, self.machine))
+            Configurer.get_reader_filenames())
         self.cat = intake.open_catalog(self.catalog_file)
 
         # check source existence
-        self.source = check_catalog_source(self.cat, self.model, self.exp, source, name="catalog")
+        self.source = check_catalog_source(self.cat, self.model, self.exp,
+                                           source, name="catalog")
 
         # check that you defined zoom in a correct way
         self.zoom = self._check_zoom(zoom)
 
         # get fixes dictionary and find them
-        self.fixes_dictionary = load_multi_yaml(self.fixer_folder)
-        self.fixes = self.find_fixes()
+        self.fix = fix
+        if self.fix:
+            self.fixes_dictionary = load_multi_yaml(self.fixer_folder)
+            self.fixes = self.find_fixes()
 
         # Store the machine-specific CDO path if available
         cfg_base = load_yaml(self.config_file)
         self.cdo = cfg_base["cdo"].get(self.machine, "cdo")
 
         # load and check the regrid
-        cfg_regrid = load_yaml(self.regrid_file)
-        source_grid_id = check_catalog_source(cfg_regrid["source_grids"],
-                                              self.model, self.exp, source, name='regrid')
-        source_grid = cfg_regrid["source_grids"][self.model][self.exp][source_grid_id]
+        if regrid or areas:
+            cfg_regrid = load_yaml(self.regrid_file)
+            source_grid_id = check_catalog_source(cfg_regrid["source_grids"],
+                                                  self.model, self.exp,
+                                                  self.source, name='regrid')
+            source_grid = cfg_regrid["source_grids"][self.model][self.exp][source_grid_id]
 
-        # Normalize vert_coord to list
-        self.vert_coord = source_grid.get("vert_coord", "2d")  # If not specified we assume that this is only a 2D case
+            # Normalize vert_coord to list
+            self.vert_coord = source_grid.get("vert_coord", "2d")  # If not specified we assume that this is only a 2D case
 
-        if not isinstance(self.vert_coord, list):
-            self.vert_coord = [self.vert_coord]
+            if not isinstance(self.vert_coord, list):
+                self.vert_coord = [self.vert_coord]
 
-        self.masked_att = source_grid.get("masked", None)  # Optional selection of masked variables
-        self.masked_vars = source_grid.get("masked_vars", None)  # Optional selection of masked variables
+            self.masked_att = source_grid.get("masked", None)  # Optional selection of masked variables
+            self.masked_vars = source_grid.get("masked_vars", None)  # Optional selection of masked variables
 
-        # Expose grid information for the source as a dictionary of open xarrays
-        sgridpath = source_grid.get("path", None)
-        if sgridpath:
-            if isinstance(sgridpath, dict):
-                self.src_grid = {}
-                for k, v in sgridpath.items():
-                    self.src_grid.update({k: xr.open_dataset(v.format(zoom=self.zoom), decode_times=False)})
-            else:
-                if self.vert_coord:
-                    self.src_grid = {self.vert_coord[0]: xr.open_dataset(sgridpath.format(zoom=self.zoom), decode_times=False)}
+            # Expose grid information for the source as a dictionary of
+            # open xarrays
+            sgridpath = source_grid.get("path", None)
+            if sgridpath:
+                if isinstance(sgridpath, dict):
+                    self.src_grid = {}
+                    for k, v in sgridpath.items():
+                        self.src_grid.update({k: xr.open_dataset(v.format(zoom=self.zoom),
+                                                                decode_times=False)})
                 else:
-                    self.src_grid = {"2d": xr.open_dataset(sgridpath.format(zoom=self.zoom), decode_times=False)}
-        else:
-            self.src_grid = None
+                    if self.vert_coord:
+                        self.src_grid = {self.vert_coord[0]: xr.open_dataset(sgridpath.format(zoom=self.zoom),
+                                                                            decode_times=False)}
+                    else:
+                        self.src_grid = {"2d": xr.open_dataset(sgridpath.format(zoom=self.zoom),
+                                                            decode_times=False)}
+            else:
+                self.src_grid = None
 
-        self.dst_datamodel = datamodel
-        # Default destination datamodel (unless specified in instantiating the Reader)
-        if not self.dst_datamodel:
-            self.dst_datamodel = self.fixes_dictionary["defaults"].get("dst_datamodel", None)
+            self.src_space_coord = source_grid.get("space_coord", None)
+            self.space_coord = self.src_space_coord
 
-        self.src_space_coord = source_grid.get("space_coord", None)
-        self.space_coord = self.src_space_coord
-        self.dst_space_coord = ["lon", "lat"]
+        if self.fix:
+            self.dst_datamodel = datamodel
+            # Default destination datamodel
+            # (unless specified in instantiating the Reader)
+            if not self.dst_datamodel:
+                self.dst_datamodel = self.fixes_dictionary["defaults"].get("dst_datamodel", None)
 
         if regrid:
+            self.dst_space_coord = ["lon", "lat"]
 
             self.weightsfile = {}
             self.weights = {}
@@ -178,16 +205,18 @@ class Reader(FixerMixin, RegridMixin):
                 # compute correct filename ending
                 levname = vc if vc == "2d" or vc == "2dm" else f"3d-{vc}"
 
-                template_file = cfg_regrid["weights"]["template"].format(model=model,
-                                                                         exp=exp,
+                template_file = cfg_regrid["weights"]["template"].format(model=self.model,
+                                                                         exp=self.exp,
                                                                          method=method,
                                                                          target=regrid,
                                                                          source=self.source,
                                                                          level=levname)
 
-                # add the zoom level in the template file (same as done in areas)
+                # add the zoom level in the template file
                 if self.zoom is not None:
-                    template_file = re.sub(r'\.nc', '_z' + str(self.zoom) + r'\g<0>', template_file)
+                    template_file = re.sub(r'\.nc',
+                                           '_z' + str(self.zoom) + r'\g<0>',
+                                           template_file)
 
                 self.weightsfile.update({vc: os.path.join(
                     cfg_regrid["weights"]["path"],
@@ -198,20 +227,26 @@ class Reader(FixerMixin, RegridMixin):
                     if os.path.exists(self.weightsfile[vc]):
                         os.unlink(self.weightsfile[vc])
                     self._make_weights_file(self.weightsfile[vc], source_grid,
-                                            cfg_regrid, regrid=regrid, vert_coord=vc,
-                                            extra=extra, zoom=self.zoom, method=method)
+                                            cfg_regrid, regrid=regrid,
+                                            vert_coord=vc, extra=extra,
+                                            zoom=self.zoom, method=method)
 
                 self.weights.update({vc: xr.open_mfdataset(self.weightsfile[vc])})
                 vc2 = None if vc == "2d" or vc == "2dm" else vc
-                self.regridder.update({vc: rg.Regridder(weights=self.weights[vc], vert_coord=vc2, space_dims=default_space_dims)})
+                self.regridder.update({vc: rg.Regridder(weights=self.weights[vc],
+                                                        vert_coord=vc2,
+                                                        space_dims=default_space_dims)})
 
         if areas:
+            template_file = cfg_regrid["areas"]["src_template"].format(model=self.model,
+                                                                       exp=self.exp,
+                                                                       source=self.source)
 
-            template_file = cfg_regrid["areas"]["src_template"].format(model=model, exp=exp, source=self.source)
-
-            # add the zoom level in the template file (same as done in weights)
+            # add the zoom level in the template file
             if self.zoom is not None:
-                template_file = re.sub(r'\.nc', '_z' + str(self.zoom) + r'\g<0>', template_file)
+                template_file = re.sub(r'\.nc',
+                                       '_z' + str(self.zoom) + r'\g<0>',
+                                       template_file)
 
             self.src_areafile = os.path.join(
                 cfg_regrid["areas"]["path"],
@@ -223,6 +258,7 @@ class Reader(FixerMixin, RegridMixin):
                 cellareas = source_grid.get("cellareas", None)
                 cellarea_var = source_grid.get("cellarea_var", None)
                 if cellareas and cellarea_var:
+                    self.logger.warning("Using cellareas file provided in regrid.yaml")
                     xr.open_mfdataset(cellareas)[cellarea_var].rename("cell_area").squeeze().to_netcdf(self.src_areafile)
                 else:
                     # We have to reconstruct it
@@ -251,7 +287,7 @@ class Reader(FixerMixin, RegridMixin):
             self.grid_area = self.src_grid_area
 
     def retrieve(self, regrid=False, timmean=False,
-                 fix=True, apply_unit_fix=True, var=None, vars=None,  # pylint: disable=W0622
+                 apply_unit_fix=True, var=None, vars=None,
                  streaming=False, stream_step=None, stream_unit=None,
                  stream_startdate=None, streaming_generator=False,
                  startdate=None, enddate=None):
@@ -259,25 +295,35 @@ class Reader(FixerMixin, RegridMixin):
         Perform a data retrieve.
 
         Arguments:
-            regrid (bool):          if to regrid the retrieved data (False)
-            timmean (bool):         if to average the retrieved data (False)
-            fix (bool):             if to perform a fix (var name, units, coord name adjustments) (True)
-            apply_unit_fix (bool):  if to already adjust units by multiplying by a factor or adding
-                                    an offset (this can also be done later with the `apply_unit_fix` method) (True)
-            var (str, list):        the variable(s) to retrieve (None), vars is a synonym
-                                    if None, all variables are retrieved
-            streaming (bool):       if to retreive data in a streaming mode (False)
-            streaming_generator (bool):  if to return a generator object for data streaming (False).
-            stream_step (int):      the number of time steps to stream the data by (Default = 1)
-            stream_unit (str):      the unit of time to stream the data by
-                                    (e.g. 'hours', 'days', 'months', 'years') (None)
-            stream_startdate (str): the starting date for streaming the data (e.g. '2020-02-25') (None)
+            regrid (bool):              if to regrid the retrieved data
+                                        Defaults to False
+            timmean (bool):             if to average the retrieved data
+                                        Defaults to False
+            apply_unit_fix (bool):      if to already adjust units by
+                                        multiplying by a factor or adding
+                                        an offset (this can also be done later
+                                        with the `apply_unit_fix` method).
+                                        Defaults to True
+            var (str, list):            the variable(s) to retrieve.
+                                        Defaults to None
+                                        vars is a synonym.
+                                        if None, all variables are retrieved
+            streaming (bool):           if to retreive data in a streaming
+                                        mode. Defaults to False
+            streaming_generator (bool): if to return a generator object for
+                                        data streaming. Defaults to False
+            stream_step (int):          the number of time steps to stream the
+                                        data by. Defaults to None
+            stream_unit (str):          the unit of time to stream the data
+                                        by (e.g. 'hours', 'days', 'months',
+                                        'years'). Defaults to None
+            stream_startdate (str):     the starting date for streaming the
+                                        data (e.g. '2020-02-25').
+                                        Defaults to None
         Returns:
             A xarray.Dataset containing the required data.
         """
 
-        # this is done in the __init__
-        # self.cat = intake.open_catalog(self.catalog_file)
         # Extract subcatalogue
         if self.zoom:
             esmcat = self.cat[self.model][self.exp][self.source](zoom=self.zoom)
@@ -289,11 +335,11 @@ class Reader(FixerMixin, RegridMixin):
 
         # get loadvar
         if var:
-            if isinstance(var, str):
+            if isinstance(var, str):  # conversion to list guarantees that a Dataset is produced
                 var = var.split()
             self.logger.info("Retrieving variables: %s", var)
 
-            loadvar = self.get_fixer_varname(var) if fix else var
+            loadvar = self.get_fixer_varname(var) if self.fix else var
         else:
             loadvar = None
 
@@ -308,7 +354,22 @@ class Reader(FixerMixin, RegridMixin):
         else:
             data = self.reader_intake(esmcat, var, loadvar)  # Returns a generator object
 
-        log_history_iter(data, "retrieved by AQUA retriever")
+            if var:
+                if all(element in data.data_vars for element in loadvar):
+                    data = data[loadvar]
+                else:
+                    try:
+                        data = data[var]
+                        self.logger.warning(f"You are asking for var {var} which is already fixed from {loadvar}.")
+                        self.logger.warning(f"Would be safer to run with fix=False")
+                    except:
+                        raise KeyError("You are asking for variables which we cannot find in the catalog!")
+
+        if fiter and self.buffer:
+            data = self.buffer_iter(data)
+            fiter = False
+
+        data = log_history_iter(data, "retrieved by AQUA retriever")
 
         # sequence which should be more efficient: decumulate - averaging - regridding - fixing
 
@@ -320,9 +381,8 @@ class Reader(FixerMixin, RegridMixin):
         if self.targetgrid and regrid:
             data = self.regrid(data)
             self.grid_area = self.dst_grid_area
-        if fix:
-            data = self.fixer(data, apply_unit_fix=apply_unit_fix)  # fixer accepts also iterators
-
+        if self.fix:   # Do not change easily this order. The fixer assumes to be after regridding
+            data = self.fixer(data, var, apply_unit_fix=apply_unit_fix)
         if not fiter:
             # This is not needed if we already have an iterator
             if streaming or self.streaming or streaming_generator:
@@ -385,23 +445,28 @@ class Reader(FixerMixin, RegridMixin):
             # If this was a single dataarray
             out = out[0]
 
-        out = set_attrs(out, {"regridded": 1})  # set regridded attribute to 1 for all vars
+        # set regridded attribute to 1 for all vars
+        out = set_attrs(out, {"regridded": 1})
 
-        # set these two to the target grid (but they are actually not used so far)
+        # set these two to the target grid
+        # (but they are actually not used so far)
         self.grid_area = self.dst_grid_area
         self.space_coord = ["lon", "lat"]
 
         log_history(out, "regridded by AQUA regridder")
         return out
 
-    def timmean(self, data, freq=None):
+    def timmean(self, data, freq=None, time_bounds=False):
         """
         Perform daily and monthly averaging
 
         Arguments:
             data (xr.Dataset):  the input xarray.Dataset
+            freq (str):         the frequency of the averaging.
+                                Defaults to None
+            time_bound (bool):  option to create the time bounds
         Returns:
-            A xarray.Dataset containing the regridded data.
+            A xarray.Dataset containing the time averaged data.
         """
 
         if freq is None:
@@ -421,19 +486,29 @@ class Reader(FixerMixin, RegridMixin):
             # resample
             self.logger.info('Resamplig to %s frequency...', str(resample_freq))
             out = data.resample(time=resample_freq).mean()
-            # for now, we set initial time of the averaging period following ECMWF standard
-            # HACK: we ignore hours/sec to uniform the output structure
-            #proper_time = data.time.resample(time=resample_freq).min()
-            #out['time'] = np.array(proper_time.values, dtype='datetime64[h]')
-            out['time'] = data.time.resample(time=resample_freq).min().dt.floor('D')
-        except ValueError:
-            sys.exit('Cant find a frequency to resample, aborting!')
+        except ValueError as exc:
+            raise ValueError('Cant find a frequency to resample, aborting!') from exc
+        
+        # set time as the first timestamp of each month/day according to the sampling frequency
+        out['time'] = out['time'].to_index().to_period(resample_freq).to_timestamp().values
 
-        # check for NaT
+        # check time is correct
         if np.any(np.isnat(out.time)):
-            self.logger.warning('Resampling cannot produce output for all frequency step, is your input data correct?')
+            raise ValueError('Resampling cannot produce output for all frequency step, is your input data correct?')
 
         log_history(out, f"resampled to frequency {resample_freq} by AQUA timmean")
+
+        # add a variable to create time_bounds
+        if time_bounds:
+            resampled = data.time.resample(time=resample_freq)
+            time_bnds = xr.concat([resampled.min(),  resampled.max()], dim='bnds').transpose()
+            time_bnds['time'] = out.time
+            time_bnds.name = 'time_bnds'
+            out = xr.merge([out, time_bnds])
+            if np.any(np.isnat(out.time_bnds)):
+                raise ValueError('Resampling cannot produce output for all time_bnds step!')
+            log_history(out, "time_bnds added by by AQUA timmean")
+       
         return out
 
     def _check_if_regridded(self, data):
@@ -453,23 +528,36 @@ class Reader(FixerMixin, RegridMixin):
 
         return att.get("regridded", False)
 
-    def fldmean(self, data):
+    def fldmean(self, data, lon_limits=None, lat_limits=None, **kwargs):
         """
         Perform a weighted global average.
+        If a subset of the data is provided, the average is performed only on the subset.
 
         Arguments:
             data (xr.DataArray or xarray.DataDataset):  the input data
+            lon_limits (list, optional):  the longitude limits of the subset
+            lat_limits (list, optional):  the latitude limits of the subset
+
+        Kwargs:
+            - box_brd (bool,opt): choose if coordinates are comprised or not in area selection.
+                                  Default is True
+
         Returns:
             the value of the averaged field
         """
 
-        # If these data have been regridded we should use the destination grid info
+        # If these data have been regridded we should use
+        # the destination grid info
         if self._check_if_regridded(data):
             space_coord = self.dst_space_coord
             grid_area = self.dst_grid_area
         else:
             space_coord = self.src_space_coord
             grid_area = self.src_grid_area
+
+        if lon_limits is not None or lat_limits is not None:
+            data = area_selection(data, lon=lon_limits, lat=lat_limits,
+                                  **kwargs)
 
         # check if coordinates are aligned
         try:
@@ -496,7 +584,7 @@ class Reader(FixerMixin, RegridMixin):
                     if self.grid_area[coord].sortby(coord).equals(xcoord.sortby(coord)):
                         self.logger.warning('%s is sorted in different way between area files and your dataset. Flipping it!', coord)
                         self.grid_area = self.grid_area.reindex({coord: list(reversed(self.grid_area[coord]))})
-                        #raise ValueError(f'{coord} is sorted in different way between area files and your dataset.') from err
+                        # raise ValueError(f'{coord} is sorted in different way between area files and your dataset.') from err
                     # something else
                     else:
                         raise ValueError(f'{coord} has a mismatch in coordinate values!') from err
@@ -506,14 +594,13 @@ class Reader(FixerMixin, RegridMixin):
         return out
 
     def _check_zoom(self, zoom):
-
         """
-        Function to check if the zoom parameter is included in the metadata of the
-        source and performs a few safety checks.
+        Function to check if the zoom parameter is included in the metadata of
+        the source and performs a few safety checks.
         It could be extended to any other metadata flag.
 
         Arguments:
-            zoom (integer):
+            zoom (int): the zoom level to be checked
 
         Returns:
             zoom after check has been processed
@@ -547,12 +634,14 @@ class Reader(FixerMixin, RegridMixin):
                                 self.model, self.exp, self.source)
             return None
 
-    def vertinterp(self, data, levels=None, vert_coord='plev', units=None, method='linear'):
+    def vertinterp(self, data, levels=None, vert_coord='plev', units=None,
+                   method='linear'):
         """
         A basic vertical interpolation based on interp function
-        of xarray within AQUA. Given an xarray object, will interpolate the vertical dimension along
-        the vert_coord. If it is a Dataset, only variables with the required vertical coordinate
-        will be interpolated
+        of xarray within AQUA. Given an xarray object, will interpolate the
+        vertical dimension along the vert_coord.
+        If it is a Dataset, only variables with the required vertical
+        coordinate will be interpolated.
 
         Args:
             data (DataArray, Dataset): your dataset
@@ -617,7 +706,7 @@ class Reader(FixerMixin, RegridMixin):
             query[query_var] = var.split() if isinstance(var, str) else var
         subcat = esmcat.search(**query)
         data = subcat.to_dataset_dict(cdf_kwargs=cdf_kwargs,
-                                      zarr_kwargs=dict(consolidated=True),
+                                      # zarr_kwargs=dict(consolidated=True),
                                       # decode_times=True,
                                       # use_cftime=True)
                                       progressbar=False
@@ -630,7 +719,16 @@ class Reader(FixerMixin, RegridMixin):
 
         if not enddate:
             enddate = startdate
-        return esmcat(startdate=startdate, enddate=enddate, var=var).read_chunked()
+
+        fdb_path = esmcat.metadata.get('fdb_path', None)
+        if fdb_path:
+            os.environ["FDB5_CONFIG_FILE"] = fdb_path
+
+        if self.aggregation:
+            return esmcat(startdate=startdate, enddate=enddate, var=var, aggregation=self.aggregation, verbose=self.verbose).read_chunked()
+        else:
+            return esmcat(startdate=startdate, enddate=enddate, var=var, verbose=self.verbose).read_chunked()
+            
 
     def reader_intake(self, esmcat, var, loadvar, keep="first"):
         """
@@ -661,9 +759,28 @@ class Reader(FixerMixin, RegridMixin):
             data = esmcat.to_dask()
 
         # check for duplicates
-        len0 = len(data.time)
-        data = data.drop_duplicates(dim='time', keep=keep)
-        if len(data.time) != len0:
-            self.logger.warning("Duplicate entries found along the time axis, keeping the %s one.", keep)
+        if 'time' in data.coords:
+            len0 = len(data.time)
+            data = data.drop_duplicates(dim='time', keep=keep)
+            if len(data.time) != len0:
+                self.logger.warning("Duplicate entries found along the time axis, keeping the %s one.", keep)
 
         return data
+
+    def buffer_iter(self, data):
+        """
+        Buffers an iterator object into a temporary directory
+        Args:
+            data (iterator over xarray.Dataset): the data to be buffered
+
+        Returns:
+            A xarray.Dataset pointing to the buffered data
+        """
+
+        self.logger.info("Buffering iterator to: %s", self.buffer.name)
+        niter =0
+        for dd in data:
+            dd.to_netcdf(f"{self.buffer.name}/iter{niter}.nc")
+            niter = niter + 1
+
+        return xr.open_mfdataset(f"{self.buffer.name}/iter*.nc")