+++ tests/test_catalog.py
@@ -5,6 +5,7 @@ import xarray
 from aqua import Reader, catalogue, inspect_catalogue
 from aqua.reader.reader_utils import check_catalog_source
 
+
 @pytest.fixture(params=[(model, exp, source)
                         for model in catalogue()
                         for exp in catalogue()[model]
@@ -22,16 +23,26 @@ def reader(request):
         pytest.skip()
     if model == 'ERA5':
         pytest.skip()
-    myread = Reader(model=model, exp=exp, source=source, areas=False)
-    data = myread.retrieve(fix=False)
+    if model == 'IFS' and source == 'fdb':  # there is another test for that
+        pytest.skip()
+    # teleconnections catalogue, only on teleconnections workflow
+    if model == 'IFS' and exp == 'test-tco79' and source == 'teleconnections':
+        pytest.skip()
+    myread = Reader(model=model, exp=exp, source=source, areas=False,
+                    fix=False)
+    data = myread.retrieve()
     return myread, data
+
 
 @pytest.mark.slow
 def test_catalogue(reader):
-    """Checking that both reader and Dataset are retrived in reasonable shape"""
+    """
+    Checking that both reader and Dataset are retrived in reasonable shape
+    """
     aaa, bbb = reader
     assert isinstance(aaa, Reader)
     assert isinstance(bbb, xarray.Dataset)
+
 
 @pytest.mark.aqua
 def test_inspect_catalogue():
@@ -43,6 +54,7 @@ def test_inspect_catalogue():
     assert isinstance(exps, list)
     sources = inspect_catalogue(cat, model='IFS', exp='test-tco79')
     assert isinstance(sources, list)
+
 
 @pytest.mark.aqua
 @pytest.mark.parametrize(
@@ -56,7 +68,8 @@ def test_inspect_catalogue():
             "source1",
             "source1"
         ),
-        # Test case 2: Source is specified but does not exist, default source exists
+        # Test case 2: Source is specified but does not exist,
+        # default source exists
         (
             {"model1": {"exp1": {"default": "default_data", "source2": "data2"}}},
             "model1",
@@ -64,7 +77,8 @@ def test_inspect_catalogue():
             "source1",
             "default"
         ),
-        # Test case 3: Source is specified but does not exist, default source does not exist
+        # Test case 3: Source is specified but does not exist,
+        # default source does not exist
         (
             {"model1": {"exp1": {"source2": "data2"}}},
             "model1",
@@ -88,7 +102,8 @@ def test_inspect_catalogue():
             None,
             pytest.raises(KeyError)
         ),
-        # Test case 6: Source is not specified, no sources available, but a default source exists
+        # Test case 6: Source is not specified, no sources available,
+        # but a default source exists
         (
             {"model1": {"exp1": {"default": "default_data"}}},
             "model1",